Auto stash before merge of "feat/shoppingCart" and "origin/feat/shoppingCart"

+++ app/Http/Controllers/BuyController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\LackofProductAmountException;
 use App\Models\BuyHistory;
 use App\Models\BuyRecord;
 use App\Models\CartItem;
 use App\Models\Product;
+use http\Message;
 use Exception;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -59,12 +61,22 @@ class BuyController extends Controller
                         'quantity' => $item->quantity,
                     ]);
                     $product = Product::where('id', $item->product->id)->first();
+                    // Check if the product amount is smaller than the item quantity
+                    if ($product->amount < $item->quantity) {
+                        throw new LackofProductAmountException();
+                    }
                     $product->update(['amount'=>$product->amount-$item->quantity]);
                     $item->delete();
                 }
             });
+        } catch(LackofProductAmountException $e){
+            $e->report();
+            return redirect()->back()->with('error', '含有庫存不足的商品！');
         } catch (\Throwable $e) {
-            return response()->noContent()->setStatusCode(400);
+            $cartItems = Auth::user()->cartItems()->orderBy('created_at')->get();
+            return view('payment', [
+                'cartItems' => $cartItems,
+            ]);
         }
         return view('bill', [
             'buyHistory' => $buyHistory,
